refactor(nika-builtin): seed the media/ core — first two re-homings
The architecture review's sequence ① begins with the two cheapest
residents, both pure moves pinned by the mock's byte-stable goldens:

media/time.rs — rfc3339_now, duplicated verbatim in image/ and tts/
(review P3.5 · the poster child of the copy set), now one seam.

media/png.rs — crc32, which PRODUCTION provenance sealing (embed.rs)
imported from the MOCK module (review P3.1 · inverted dependency);
integrity math now lives where both the mock and the embedder draw it,
with the PNG test vectors pinned (0xCBF43926 · the IEND constant).

The module doc states the growth rule: re-homing only, never
speculation; sub-module boundaries align with the reserved
nika-media-* crate split so graduation stays a git mv. 235 builtin
tests · clippy 0 · ratchets green.
