+++ crates/nika-builtin/src/lib.rs
@@ -34,6 +34,7 @@ pub mod defs;
 pub mod file;
 pub mod image;
 pub mod inspect;
+pub(crate) mod media;
 pub mod net;
 pub mod permits;
 pub mod tts;
